refactor(ocr): extract execution-provider policy into engine/providers
